발사모드(Single/FullAuto/Burst)·HP/Mana 회복·ADS 조준 시스템 추가
- MainWeaponComponent: FireMode(Single/FullAuto/Burst) 실제 구현
  - FullAuto는 클라이언트 반복 타이머, Burst는 서버 타이머로 나머지 탄 처리(매 탄마다 실시간 조준 방향 재조회)
  - EquipWeapon() 무기 교체 시 이전 무기의 FullAuto/Burst 타이머 정리, WeaponIndex 복제 및 OnRep로 클라이언트 재로드
  - StartADS/StopADS/Server_SetAiming으로 조준 상태 추가, GetADSSpeed() 노출
  - 클라이언트에서도 EquipWeapon()이 실행되도록 BeginPlay의 HasAuthority 제한 제거 (weapons.json 로컬 로딩)
  - GetWorld() null 체크 누락으로 인한 StopFire() 크래시 수정
- MainHPComponent/MainManaComponent: MainEffectTickComponent 신설
  - 주기적 회복 신호(OnEffectTick)를 별도 컴포넌트가 브로드캐스트, HP/Mana가 각자 구독해서 반응
  - HP는 HPRegen/HPRegenDelay(피격 후 회복 재개 지연) 반영, PlayerBaseStat.h의 MaxHP 필드명 변경 반영
- WeaponDataManager: static 멤버 정의 누락으로 인한 LNK2001 링크 에러 및 PIE 멀티플레이어 콘솔 명령어 중복 등록 크래시 수정
- BP_MainCharacter: 1인칭 카메라/무기 뷰모델(FirstPersonCamera, FirstPersonMesh) 및 ADS 시 조준경 정렬 연출 추가
- MainHUDWidget 신규: 본인 Hp/Mana/Ammo 표시 HUD, 사망 시 자동 제거

+++ Content/Data/player_base_stat_converter.py
@@ -70,12 +70,12 @@ class FieldSpec:
 
 
 PLAYER_STAT_SCHEMA: list[FieldSpec] = [
-    FieldSpec(name="MaxHP", type=int, required=True, min_value=1),
+    FieldSpec(name="MaxHP", type=float, required=True, min_value=1.0),
     FieldSpec(name="HPRegen", type=float, required=True, min_value=0.0),
     FieldSpec(name="HPRegenDelay", type=float, required=True, min_value=0.0),
     FieldSpec(name="WalkSpeed", type=float, required=True, min_value=0.0),
     FieldSpec(name="RunSpeed", type=float, required=True, min_value=0.0),
-    FieldSpec(name="MaxMana", type=int, required=True, min_value=0),
+    FieldSpec(name="MaxMana", type=float, required=True, min_value=0.0),
     FieldSpec(name="ManaRegen", type=float, required=True, min_value=0.0),
     FieldSpec(name="JumpPower", type=float, required=True, min_value=0.0),
     FieldSpec(name="Comment", type=str, required=False, export=False),
@@ -83,12 +83,12 @@ PLAYER_STAT_SCHEMA: list[FieldSpec] = [
 
 # xlsx 템플릿용 표시 정보 (한글 설명, 예시값). PLAYER_STAT_SCHEMA와 순서/이름이 1:1로 맞아야 함.
 TEMPLATE_COLUMNS = [
-    {"key": "MaxHP",        "header_kr": "최대 체력",                "type": "int",    "example": 100},
+    {"key": "MaxHP",        "header_kr": "최대 체력",                "type": "float",  "example": 100.0},
     {"key": "HPRegen",      "header_kr": "초당 체력 회복량",          "type": "float",  "example": 2.0},
     {"key": "HPRegenDelay", "header_kr": "피격 후 체력회복 지연시간(초)", "type": "float", "example": 5.0},
     {"key": "WalkSpeed",    "header_kr": "걷기 이동속도",             "type": "float",  "example": 300.0},
     {"key": "RunSpeed",     "header_kr": "달리기 이동속도",           "type": "float",  "example": 600.0},
-    {"key": "MaxMana",      "header_kr": "최대 마나",                 "type": "int",    "example": 1000},
+    {"key": "MaxMana",      "header_kr": "최대 마나",                 "type": "float",  "example": 1000.0},
     {"key": "ManaRegen",    "header_kr": "초당 마나 회복량",          "type": "float",  "example": 10.0},
     {"key": "JumpPower",    "header_kr": "점프력",                    "type": "float",  "example": 420.0},
     {"key": "Comment",      "header_kr": "기획 메모(변환시 제외)",   "type": "string", "example": "기본 스탯"},
